refactor: extract fee tracking types to fees.rs
Move FeeRateAndMinimum, NetworkFeeTracker, and BrokerFeesTracker out of
the monolithic lib.rs into a dedicated fees.rs module. Fields that were
previously private to the pallet module are made pub(crate) to maintain
access from migrations and sibling modules.
